task(NOC-1913): updated remaining references from "paragraphs" to "chunks"
