+++ server/libs/services/documents/context/telnyx.ts
@@ -1,18 +1,16 @@
 import PDFParser from 'pdf2json';
 
 import { AxiosInstance } from 'axios';
-import { Indexes, Match, PromptType, RawMatch } from '../../../../types/classes/context.js';
 import { Context } from './context.js';
 import { encode } from 'gpt-3-encoder';
 import { MarkdownSplitter } from '../splitter/markdown.js';
 import { IntercomSplitter } from '../splitter/intercom.js';
-import { Paragraph } from '../../../../types/classes/loader.js';
 import { us_central_1 } from '../../../../clients/telnyx.js';
 import { DownstreamError } from './errors.js';
-import { TelnyxLoaderType } from '../vectorstore/telnyx.js';
 import { UnstructuredTextSplitter } from '../splitter/unstructured.js';
 import { JSONSplitter } from '../splitter/json.js';
 import { PDFSplitter } from '../splitter/pdf.js';
+import { Indexes, Match, RawMatch, TelnyxBucketChunk, TelnyxContextResult, TelnyxLoaderType } from '../types.js';
 
 type DocumentSplitter = {
   identifier: string;
@@ -21,7 +19,7 @@ type DocumentSplitter = {
   display_title?: string;
   description: string;
   body: string;
-  paragraphs: Paragraph[];
+  chunks: TelnyxBucketChunk[];
   override?: string;
   total_tokens: number;
 };
@@ -39,9 +37,10 @@ export class TelnyxContext extends Context {
    * @param matches An array of vectorstore matches
    * @param max_tokens Prevents the returned string from exceeding the token limit set
    * @returns A string of documents, readable by the LLM
+   * @throws DownstreamError
    */
 
-  public async prompt(matches: Match[], max_tokens = this.MAX_DOCUMENT_TOKENS): Promise<PromptType> {
+  public async prompt(matches: Match[], max_tokens = this.MAX_DOCUMENT_TOKENS): Promise<TelnyxContextResult> {
     if (matches.length === 0) return { context: '', used: [] };
 
     try {
@@ -60,30 +59,101 @@ export class TelnyxContext extends Context {
       const detail = 'Failed to generate the context.';
       const message = e?.message;
 
-      return { error: DownstreamError(detail, message) };
+      const msg = DownstreamError(detail, message);
+      throw new Error(JSON.stringify(msg));
     }
+  }
+
+  /**
+   * Converts a string input into an array of matches using vectorstore similiarity search
+   * @param query The input to convert
+   * @param indexes The indexes to search, all if not specified
+   * @returns An array of matches
+   */
+
+  public async matches(
+    query: string,
+    indexes: Indexes[] = [],
+    max_results?: number,
+    min_certainty: number = this.MIN_CERTAINTY
+  ): Promise<Match[]> {
+    try {
+      const vector_start = performance.now();
+      const raw_matches = await this.vectorstore.query(query, indexes, max_results);
+      const vector_end = performance.now();
+
+      if (this.callback)
+        this.callback({
+          type: 'timer',
+          value: { name: 'Vector Query', duration: (vector_end - vector_start) / 1000 },
+        });
+
+      const enrich_start = performance.now();
+      const matches = await this.enrich(raw_matches, min_certainty);
+      const enrich_end = performance.now();
+
+      if (this.callback) {
+        this.callback({
+          type: 'timer',
+          value: { name: 'Enrich Results', duration: (enrich_end - enrich_start) / 1000 },
+        });
+
+        // send matches to the frontend over the SSE connection
+        this.callback({
+          type: 'matches',
+          value: matches.map((match) => ({
+            title: match.title,
+            url: match.url,
+            type: match.type,
+            certainty: match.matched.chunk.certainty,
+            tokens: match.total_tokens,
+          })),
+        });
+      }
+
+      return matches;
+    } catch (e) {
+      console.log(e);
+      throw new Error(e?.message || 'An unexpected error occured');
+    }
+  }
+
+  /**
+   * Convert a vectorstore match into the original document
+   * @param raw_matches A list of matches
+   * @returns A list of original documents
+   */
+
+  private async enrich(raw_matches: RawMatch[], min_certainty: number): Promise<Match[]> {
+    const matches = raw_matches.filter((match) => match.chunk.certainty > min_certainty);
+    if (matches.length === 0) return [];
+
+    const documents = await this.matchesToDocuments(matches);
+
+    // sort documents by certainty
+    return documents.sort((a, b) => b.matched.certainty - a.matched.certainty);
   }
 
   public async raw_matches(query: string, indexes: Indexes[] = undefined, max_results?: number) {
     return this.vectorstore.query(query, indexes, max_results);
   }
 
-  public async describe(matches: Match[]): Promise<PromptType> {
+  public async describe(matches: Match[]): Promise<TelnyxContextResult> {
     if (matches.length === 0) return { context: '', used: [] };
 
     const descriptions = [];
     const used = [];
 
     for (const match of matches) {
-      const { identifier, title, description, url, paragraphs } = match;
+      const { identifier, title, description, url, chunks } = match;
       const metadata = [`# Document ID: ${identifier}`, `- Type: \`telnyx\``];
 
       if (title) metadata.push(`- Title: ${title}`);
       if (description) metadata.push(`- Description: ${description.replaceAll('\n', ' ')}`);
       if (url) metadata.push(`- URL: ${url}`);
-      if (paragraphs.length) {
+      if (chunks.length) {
         metadata.push(
-          `- Paragraph Headings: [${paragraphs
+          `- Paragraph Headings: [${chunks
             .map((para) => `"${para.heading}"`)
             .slice(0, 20)
             .join(', ')}]`
@@ -119,7 +189,7 @@ export class TelnyxContext extends Context {
       url: item.url,
       title: item.title,
       description: item.description,
-      paragraphs: item.paragraphs,
+      chunks: item.chunks,
       total_tokens: item.total_tokens,
       override: item.override,
       loader_type: item.match.loader_type,
@@ -133,7 +203,7 @@ export class TelnyxContext extends Context {
    * @returns The full document that was matched
    */
 
-  public async matchToDocument(match: RawMatch): Promise<PromptType> {
+  public async matchToDocument(match: RawMatch): Promise<TelnyxContextResult> {
     const bucket = match.chunk.bucket;
     const document_id = match.identifier;
 
@@ -174,7 +244,7 @@ export class TelnyxContext extends Context {
       const { url, title, description, body } = this.formatSEO(document);
 
       const splitter = new MarkdownSplitter({ file: { body: body } });
-      const paragraphs = splitter.split();
+      const chunks = splitter.split();
 
       return {
         identifier: document_id,
@@ -182,8 +252,8 @@ export class TelnyxContext extends Context {
         title,
         description,
         body: body,
-        paragraphs: paragraphs,
-        total_tokens: paragraphs.reduce((total, paragraph) => total + paragraph.tokens, 0),
+        chunks: chunks,
+        total_tokens: chunks.reduce((total, chunk) => total + chunk.tokens, 0),
       };
     }
 
@@ -191,7 +261,7 @@ export class TelnyxContext extends Context {
       const { title, description, url, body } = document;
 
       const splitter = new IntercomSplitter({ article: document });
-      const paragraphs = splitter.split();
+      const chunks = splitter.split();
 
       return {
         identifier: url,
@@ -199,14 +269,14 @@ export class TelnyxContext extends Context {
         title,
         description,
         body,
-        paragraphs: paragraphs,
-        total_tokens: paragraphs.reduce((total, paragraph) => total + paragraph.tokens, 0),
+        chunks: chunks,
+        total_tokens: chunks.reduce((total, chunk) => total + chunk.tokens, 0),
       };
     }
 
     if (loader_type === TelnyxLoaderType.UnstructuredText) {
       const splitter = new UnstructuredTextSplitter({ file: document });
-      const paragraphs = splitter.split();
+      const chunks = splitter.split();
 
       return {
         identifier: document_id,
@@ -214,14 +284,14 @@ export class TelnyxContext extends Context {
         title: null,
         description: null,
         body: document,
-        paragraphs: paragraphs,
-        total_tokens: paragraphs.reduce((total, paragraph) => total + paragraph.tokens, 0),
+        chunks: chunks,
+        total_tokens: chunks.reduce((total, chunk) => total + chunk.tokens, 0),
       };
     }
 
     if (loader_type === TelnyxLoaderType.PDF) {
       const splitter = new PDFSplitter({ file: document });
-      const paragraphs = splitter.split();
+      const chunks = splitter.split();
 
       return {
         identifier: document_id,
@@ -229,14 +299,14 @@ export class TelnyxContext extends Context {
         title: null,
         description: null,
         body: document,
-        paragraphs: paragraphs,
-        total_tokens: paragraphs.reduce((total, paragraph) => total + paragraph.tokens, 0),
+        chunks: chunks,
+        total_tokens: chunks.reduce((total, chunk) => total + chunk.tokens, 0),
       };
     }
 
     if (loader_type === TelnyxLoaderType.JSON) {
       const splitter = new JSONSplitter({ file: document });
-      const paragraphs = await splitter.split();
+      const chunks = await splitter.split();
 
       return {
         identifier: document_id,
@@ -244,14 +314,14 @@ export class TelnyxContext extends Context {
         title: null,
         description: null,
         body: document,
-        paragraphs: paragraphs,
-        total_tokens: paragraphs.reduce((total, paragraph) => total + paragraph.tokens, 0),
+        chunks: chunks,
+        total_tokens: chunks.reduce((total, chunk) => total + chunk.tokens, 0),
       };
     }
 
     if (loader_type === TelnyxLoaderType.CSV) {
       const splitter = new JSONSplitter({ file: document });
-      const paragraphs = await splitter.split(true);
+      const chunks = await splitter.split(true);
 
       return {
         identifier: document_id,
@@ -259,8 +329,8 @@ export class TelnyxContext extends Context {
         title: null,
         description: null,
         body: document,
-        paragraphs: paragraphs,
-        total_tokens: paragraphs.reduce((total, paragraph) => total + paragraph.tokens, 0),
+        chunks: chunks,
+        total_tokens: chunks.reduce((total, chunk) => total + chunk.tokens, 0),
       };
     }
 
@@ -270,24 +340,24 @@ export class TelnyxContext extends Context {
       title: null,
       description: null,
       body: document,
-      paragraphs: [],
+      chunks: [],
       total_tokens: 0,
     };
   };
 
   private formatTelnyxDocument(data): string {
-    const { identifier, url, title, description, paragraphs } = data;
+    const { identifier, url, title, description, chunks } = data;
 
     let output = `# Document ID: ${identifier}\n- Title: ${title}\n- Description: ${description}\n- URL: ${url}\n\n`;
 
-    for (const paragraph of paragraphs) {
-      output += `### ${paragraph.heading}\n${paragraph.content}\n\n`;
+    for (const chunk of chunks) {
+      output += `### ${chunk.heading}\n${chunk.content}\n\n`;
     }
 
     return output;
   }
 
-  private async pdfToDocument(match: RawMatch): Promise<PromptType> {
+  private async pdfToDocument(match: RawMatch): Promise<TelnyxContextResult> {
     const bucket = match.chunk.bucket;
     const document_id = match.identifier;
     const url = `/${bucket}/${encodeURIComponent(document_id)}`;
@@ -334,10 +404,10 @@ export class TelnyxContext extends Context {
       const keywords = seo.match(keywords_regex);
 
       return {
-        title: title?.[0]?.replace('title: ', '').replaceAll('\n', '') || 'No Title',
-        description: description?.[0]?.replace('description: ', '').replaceAll('\n', '') || 'No Description',
-        url: url?.[0]?.replace('url: ', '').replaceAll('\n', '') || 'No URL Found',
-        keywords: keywords?.[0]?.replace('keywords: ', '').replaceAll('\n', '') || 'No Keywords',
+        title: title?.[0]?.replace('title: ', '').replace(/\\|\"|\'|\n/g, '') || 'No Title',
+        description: description?.[0]?.replace('description: ', '').replace(/\\|\"|\'|\n/g, '') || 'No Description',
+        url: url?.[0]?.replace('url: ', '').replace(/\\|\"|\'|\n/g, '') || 'No URL Found',
+        keywords: keywords?.[0]?.replace('keywords: ', '').replace(/\\|\"|\'|\n/g, '') || 'No Keywords',
         body: body.replace(seo_regex, ''),
       };
     } catch (e) {
