--- server/types/classes/chatbot.ts
@@ -1,6 +1,0 @@
-export type ChatbotQuestion = {
-  user_id: string;
-  session_id?: string | null;
-  message_id: string;
-  question: string;
-};
